Remove redundant casts.

+++ MetadataExtractor.Tests/Formats/Iptc/Iso2022ConverterTest.cs
@@ -10,7 +10,7 @@ namespace MetadataExtractor.Tests.Formats.Iptc
         public void TestConvertIso2022CharsetToJavaCharset()
         {
             Assert.AreEqual("UTF-8", Iso2022Converter.ConvertIso2022CharsetToJavaCharset(new byte[] { unchecked(0x1B), unchecked(0x25), unchecked(0x47) }));
-            Assert.AreEqual("ISO-8859-1", Iso2022Converter.ConvertIso2022CharsetToJavaCharset(new byte[] { unchecked(0x1B), unchecked((byte)0xE2), unchecked((byte)0x80), unchecked((byte)0xA2), unchecked(0x41) }));
+            Assert.AreEqual("ISO-8859-1", Iso2022Converter.ConvertIso2022CharsetToJavaCharset(new byte[] { unchecked(0x1B), unchecked(0xE2), unchecked(0x80), unchecked(0xA2), unchecked(0x41) }));
         }
     }
 }
